requirements.txt
Inflector
backports.cached_property
Pillow
requests
retrying
ruamel.yaml
testfixtures
parse
pytest
loguru
notifiers